ljwx-bigscreen-new/requirements.txt
# Flask核心框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3

# 数据库驱动
PyMySQL==1.1.0
redis==5.0.1

# 日志和监控
python-dotenv==1.0.0

# 工具库
requests==2.31.0
pytz==2023.3

# 开发工具(可选)
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2 